Reorganize React definitions
