+++ react/README.md
@@ -1,11 +1,5 @@
 # React v0.13.3 Type Definitions
 
 This folder contains the following `.d.ts` files:
-* `react.d.ts` declares the external module `"react"`
-* `react-addons.d.ts` declares the external module `"react/addons"`
-* `react-global.d.ts` declares the internal module `React` in the global namespace
-* `react-addons-global.d.ts` extends the global `React` module with `addons`
-
-Interfaces are duplicated between these files; please take care to keep them in sync when making changes.
-See [#3615](https://github.com/borisyankov/DefinitelyTyped/pull/3615) for relevant discussion.
-
+* `react.d.ts` declares the module `"react"` and `"react/addons"`
+* `react-namespace.d.ts` declares the global namespace `React` (only include this if you are actually using the global `React`)
