+++ react/react-addons-tests.ts
@@ -1,4 +1,4 @@
-/// <reference path="react-addons.d.ts" />
+/// <reference path="react.d.ts" />
 import React = require("react/addons");
 
 import TestUtils = React.addons.TestUtils;
@@ -206,16 +206,15 @@ myComponent.reset();
 // --------------------------------------------------------------------------
 
 var children: any[] = ["Hello world", [null], React.DOM.span(null)];
-var divStyle = { // CSSProperties
-    flex: "1 1 main-size",
-    backgroundImage: "url('hello.png')"
-};
 var htmlAttr: React.HTMLAttributes = {
     key: 36,
     ref: "htmlComponent",
     children: children,
     className: "test-attr",
-    style: divStyle,
+    style: { // CSSProperties
+        flex: "1 1 main-size",
+        backgroundImage: "url('hello.png')"
+    },
     onClick: (event: React.MouseEvent) => {
         event.preventDefault();
         event.stopPropagation();
